gluonts code deduplicated
