+++ sktime/datatypes/_panel/_convert.py
@@ -1152,54 +1152,6 @@ if _check_soft_dependencies("polars", severity="none"):
     )
 
 if _check_soft_dependencies("gluonts", severity="none"):
-    from sktime.datatypes._adapter.gluonts import (
-        convert_listDataset_to_pandas,
-        convert_pandas_multiindex_to_pandasDataset,
-        convert_pandas_to_listDataset,
-        convert_pandasDataset_to_pandas,
-    )
+    from sktime.datatypes._adapter.gluonts import series_and_panels_gluonts_conversions
 
-    # Utilizing functions defined in _adapter/gluonts.py
-    def convert_gluonts_listDataset_to_pandas(obj, store=None):
-        return convert_listDataset_to_pandas(obj)
-
-    def convert_pandas_to_gluonts_listDataset(obj, store=None):
-        return convert_pandas_to_listDataset(obj)
-
-    def convert_pandas_multiindex_to_gluonts_pandasDataset(obj, store=None):
-        return convert_pandas_multiindex_to_pandasDataset(obj)
-
-    def convert_gluonts_pandasDataset_to_pandas_multiindex(obj, store=None):
-        return convert_pandasDataset_to_pandas(obj)
-
-    # Storing functions in convert_dict
-    convert_dict[("pd-multiindex", "gluonts_ListDataset_panel", "Panel")] = (
-        convert_pandas_to_gluonts_listDataset
-    )
-
-    convert_dict[("gluonts_ListDataset_panel", "pd-multiindex", "Panel")] = (
-        convert_gluonts_listDataset_to_pandas
-    )
-
-    convert_dict[("pd-multiindex", "gluonts_PandasDataset_panel", "Panel")] = (
-        convert_pandas_multiindex_to_gluonts_pandasDataset
-    )
-
-    convert_dict[("gluonts_PandasDataset_panel", "pd-multiindex", "Panel")] = (
-        convert_gluonts_pandasDataset_to_pandas_multiindex
-    )
-
-    # Extending conversions
-    _extend_conversions(
-        "gluonts_ListDataset_panel",
-        "pd-multiindex",
-        convert_dict,
-        mtype_universe=MTYPE_LIST_PANEL,
-    )
-
-    _extend_conversions(
-        "gluonts_PandasDataset_panel",
-        "pd-multiindex",
-        convert_dict,
-        mtype_universe=MTYPE_LIST_PANEL,
-    )
+    convert_dict = series_and_panels_gluonts_conversions(convert_dict)
