+++ sktime/datatypes/_series/_convert.py
@@ -318,54 +318,6 @@ if _check_soft_dependencies("polars", severity="none"):
 
 
 if _check_soft_dependencies("gluonts", severity="none"):
-    from sktime.datatypes._adapter.gluonts import (
-        convert_listDataset_to_pandas,
-        convert_pandas_dataframe_to_pandasDataset,
-        convert_pandas_to_listDataset,
-        convert_pandasDataset_to_pandas_dataframe,
-    )
+    from sktime.datatypes._adapter.gluonts import series_and_panels_gluonts_conversions
 
-    # Utilizing functions defined in _adapter/gluonts.py
-    def convert_gluonts_listDataset_to_pandas(obj, store=None):
-        return convert_listDataset_to_pandas(obj)
-
-    def convert_pandas_to_gluonts_listDataset(obj, store=None):
-        return convert_pandas_to_listDataset(obj)
-
-    def convert_gluonts_PandasDataset_to_pandas(obj, store=None):
-        return convert_pandasDataset_to_pandas_dataframe(obj)
-
-    def convert_pandas_to_gluonts_PandasDataset(obj, store=None):
-        return convert_pandas_dataframe_to_pandasDataset(obj)
-
-    # Storing functions in convert_dict
-    convert_dict[("pd.DataFrame", "gluonts_ListDataset_series", "Series")] = (
-        convert_pandas_to_gluonts_listDataset
-    )
-
-    convert_dict[("gluonts_ListDataset_series", "pd.DataFrame", "Series")] = (
-        convert_gluonts_listDataset_to_pandas
-    )
-
-    convert_dict[("pd.DataFrame", "gluonts_PandasDataset_series", "Series")] = (
-        convert_pandas_to_gluonts_PandasDataset
-    )
-
-    convert_dict[("gluonts_PandasDataset_series", "pd.DataFrame", "Series")] = (
-        convert_gluonts_PandasDataset_to_pandas
-    )
-
-    # Extending conversions
-    _extend_conversions(
-        "gluonts_ListDataset_series",
-        "pd.DataFrame",
-        convert_dict,
-        mtype_universe=MTYPE_LIST_SERIES,
-    )
-
-    _extend_conversions(
-        "gluonts_PandasDataset_series",
-        "pd.DataFrame",
-        convert_dict,
-        mtype_universe=MTYPE_LIST_SERIES,
-    )
+    convert_dict = series_and_panels_gluonts_conversions(convert_dict)
